untaint $uid.
Found out that $uid is in fact tainted while trying to use it in a
system() call in another context.  We have untainted it here.

I think this actually means that perl's taint checking is broken,
because it fails to check for taint in the open() call to a
subprocesses.