Add client module to query agent
A new command-line utility, msva-query-agent, is created that can
query an msva.  It takes the pkc data on stdin, and then the context,
peer, and pkc type as arguments.

An alternate command, msva-review-cert, is also included that actually
uses the msva library to review the cert directly, without needing to
query an agent over a network socket.